indeksiä

+++ harjoitustyo/src/entities/all_ingredients.py
@@ -16,7 +16,7 @@ class Hop:
 
 
 class Yeast:
-    def __init__(self, name: str, attenuation: int, temperature_range: str):
+    def __init__(self, name: str, attenuation: float, temperature_range: str):
         self.name = name
         self.attenuation = attenuation
         self.temperature_range = temperature_range
